Add config file parsing and add option to set default encoding frame size.
Config file is in /etc/enigma2/streamproxy.sh, options are like:
listen = 8002:transcode

Both command line and config file options are honoured, command
line options take precedence.

The new option --size sets the default frame size. In a future
commit, a client will be able to request a frame size on a per
session base, but first implement a default value. The value can
be either 480p, 576p and 720p. This can be given as "--size ..."
or it can be set in the config file as "size = ....".