Sigh... Workarounds for fatal driver bugs in the vusolo2.
It appears that you cannot write, READ or access ANY
proc-file on the solo2 under /proc/encoder/0 while the
encoder is open, if you do, your userspace program will
SEGFAULT.

To make matters worse, if you try to check the presence
of /dev/bcm_enc1 (which isn't there on the solo2), to
prepare for a workaround, and you do that BEFORE or AFTER
you open the encoder, your program will SEGFAULT as well.

So now made a workoround using readdir, which seems to work.