Make file streaming/transcoding start quicker.
Normally the first and last timestamps (PCR) are retrieved
from the stream to be able to search quickly for the starting
point ("startfrom"). This may be a lengthy process on some
streams, though.

Now when no startfrom parameter is given (which currently most
always is the case), simply skip seeking for the PCR values.