Consider scaling factor when validating parameters.
E.g. for bitrate, first try raw value, i.e. value in kbps. Some apps
supply the bitrate in bps, so also try the "scaled" value if it is
deemed to be invalid, in case of bitrate, the scaling value is 1000,
so assume bps instead of kbps and try to validate that value.